Rename `RequiredCourseList` to `ProgramRequirementsList`
