+++ site/src/pages/RoadmapPage/sidebar/ProgramRequirementsList.scss
